requirements.txt
Django>=3.0.14,<3.1.14
djangorestframework>=3.10.0,<3.12.4

flake8>=3.9.0,<4.0.0